Move validator mocha tests to jest
